stocks_parser/requirements.txt
tabula-py==2.1.0
XlsxWriter==1.2.8
